Remove template rwx from adminapi role
This doesn't really add anything for me right now. It allows a mgmtvm to
install, update, and remove template VMs. This might be useful if your
playbooks depend on non-default templates or you really need automatic
template updating with Ansible. But this happens so seldom that I would
rather use the template management GUI provided in dom0 to handle these
updates manually.

This policy depended on using the qubes_prefs lookup plugin to find the
system's updatevm. If the mgmtvm hadn't actually granted itself global-ro
by the time this task runs, the lookup will fail (either with an
exception or silently and return an emtpy string, causing the policy lint
to fail) and terminate the rest of the play. This could happen when using
--check mode, where the tasks would show what they would do, including
adding the adminvm to include/global-ro, without actually making the
change.